feat: add driver check to trip creation, uses random from db

+++ src/domain/features/trip/create-trip.usecase.ts
@@ -1,23 +1,27 @@
 import type { CreateTripSchema } from "@/adapters/api/validation-schemas/trip.schema";
+import type { Driver } from "@/domain/features/driver/driver.model";
 import type { Fare } from "@/domain/features/fare/fare.model";
 import type { FareRepository } from "@/domain/features/fare/fare.repository";
 import type { Trip } from "@/domain/features/trip/trip.model";
 import type { TripNotifier } from "@/domain/features/trip/trip.notifier";
-import type { PassengerRepository } from "@/domain/shared/base-user.repository";
+import type { DriverRepository, PassengerRepository } from "@/domain/shared/base-user.repository";
 import { AppError } from "@/domain/utils/app-error";
 import { type AsyncResult, R } from "@/domain/utils/result";
 
 export class CreateTripUseCase {
   private readonly passengerRepository: PassengerRepository;
+  private readonly driverRepository: DriverRepository;
   private readonly fareRepository: FareRepository;
   private readonly tripNotifier: TripNotifier;
 
   constructor(
     passengerRepository: PassengerRepository,
+    driverRepository: DriverRepository,
     fareRepository: FareRepository,
     tripNotifier: TripNotifier,
   ) {
     this.passengerRepository = passengerRepository;
+    this.driverRepository = driverRepository;
     this.fareRepository = fareRepository;
     this.tripNotifier = tripNotifier;
   }
@@ -43,16 +47,25 @@ export class CreateTripUseCase {
       );
     }
 
-    const trip = this.createReceipt(request, fare.value);
+    const driver = await this.driverRepository.getNearest();
+    if (driver.isError) {
+      return R.error(driver.error);
+    }
+    if (driver.value === null) {
+      return R.error(new AppError("noDriversAvailable", "no drivers available to fulfill request"));
+    }
+
+    const trip = this.createReceipt(request, fare.value, driver.value);
     this.fareRepository.delete(fare.value.requestId);
     this.tripNotifier.notify(trip);
 
     return R.ok(trip);
   }
 
-  private createReceipt(request: CreateTripSchema, fare: Fare): Trip {
+  private createReceipt(request: CreateTripSchema, fare: Fare, driver: Driver): Trip {
     return {
       passengerId: request.passengerId,
+      driverId: driver.id,
       datetime: fare.datetime,
       distanceInKm: fare.distanceInKm,
       price: fare.price,
